+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
  - **Breaking:** The service definition yaml format was changed to directly specify policies and tasks instead of desired resilience levels.
  - **Breaking:** Adapted the `steadybit service verify` behavior so that it is compatible with the reworked API model.
+ - `steadybit service apply` now helps to resolve conflicts.
 
 ## v1.0.0
  - First public release.